Code up basic dumb AI. Make a nice macro
